feat (ugc): moves common files to utils
