+++ backend/apps/ugc_api/src/utils/base_schema.py
@@ -1,0 +1,12 @@
+import orjson
+from pydantic import BaseModel
+
+
+def orjson_dumps(value, *, default):
+    return orjson.dumps(value, default=default).decode()
+
+
+class BaseSchema(BaseModel):
+    class Config:
+        json_loads = orjson.loads
+        json_dumps = orjson_dumps
